MBA - final project - Data Science & Analytics
Abstract

The city of São Paulo is known for its contrasts between rich and poor neighborhoods, and inequality is a chronic issue affecting the lives of millions of people. It is essential to highlight that inequality also has racial aspects, as the Black population has been historically disadvantaged and continues to face barriers in accessing basic goods and services. Obtaining data to understand the magnitude and nature of this problem is crucial. However, dealing with a large volume of information is not simple, as the data are complex and difficult to interpret. Additionally, the lack of standardization in data collection and presentation can make analysis even more challenging. In this context, unsupervised statistical techniques such as PCA (Principal Component Analysis) and Cluster Analysis can be used to perform tasks such as dimensionality reduction, pattern and trend identification in the data, and visualization of groups or clusters of similar observations. The objective is to demonstrate how these techniques can be applied to analyze social and economic data and identify patterns and trends that can be used to understand and address inequality in the city. Finally, it is concluded that the combination of data analysis and statistical techniques can be an effective tool in tackling complex social challenges, aiding in the creation of public policies to combat socioeconomic inequality.

Keywords: Unsupervised learning; Principal component analysis; Cluster analysis; HDI; Racism.